+++ webview-sys/lib.rs
@@ -42,6 +42,6 @@ extern "C" {
     pub fn webview_set_minimized(this: *mut CWebView, minimize: c_int);
     pub fn webview_set_visible(this: *mut CWebView, visible: c_int);
     pub fn webview_set_color(this: *mut CWebView, red: u8, green: u8, blue: u8, alpha: u8);
-    pub fn webview_set_zoom_level(this: *mut CWebView, percentage: c_float);
+    pub fn webview_set_zoom_level(this: *mut CWebView, percentage: c_double);
     pub fn webview_set_html(this: *mut CWebView, html: *const c_char);
 }
